refactor: move main script and extract helpers
