+++ index.html
@@ -40,6 +40,6 @@
   </div>
   <div id="drop-container">載入中...</div>
 
-  <script src="main.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
